Migrate email address to config file
This makes more sense than specifying it as an environment variable, now
that the script is more complex.

To support the switch between sending an email and printing the report
text, command line arguments are now supported when main.py is run as a
script. The AWS Lambda handler always sends an email via SES.